utils.cpp/h : Refactor for reduced redundancy
