+++ cpp/libs/CMakeLists.txt
@@ -2,9 +2,7 @@ add_subdirectory(bisect_expected)
 add_subdirectory(bisect_nmoscpp)
 add_subdirectory(bisect_sdp)
 add_subdirectory(bisect_gst)
-add_subdirectory(gst_nmos_sender_plugin)
-add_subdirectory(gst_nmos_video_receiver_plugin)
-add_subdirectory(gst_nmos_audio_receiver_plugin)
+add_subdirectory(gst_nmos_plugins)
 add_subdirectory(ossrf_nmos_api)
 add_subdirectory(ossrf_gstreamer_api)
 add_subdirectory(bisect_json)
