+++ cpp/libs/gst_nmos_plugins/src/gst_nmos_audio_receiver_plugin.cpp
@@ -27,22 +27,24 @@
 #include "bisect/nmoscpp/configuration.h"
 #include "utils.h"
 #include "ossrf/nmos/api/nmos_client.h"
+#include "../include/element_class.h"
+#include "../include/nmos_configuration.h"
 #include <gst/gst.h>
 #include <gst/gstpad.h>
 
 using namespace bisect;
 using namespace bisect::sdp;
 
-GST_DEBUG_CATEGORY_STATIC(gst_nmosvideoreceiver_debug_category);
-#define GST_CAT_DEFAULT gst_nmosvideoreceiver_debug_category
-#define GST_TYPE_NMOSVIDEORECEIVER (gst_nmosvideoreceiver_get_type())
-#define GST_NMOSVIDEORECEIVER(obj) (G_TYPE_CHECK_INSTANCE_CAST((obj), GST_TYPE_NMOSVIDEORECEIVER, GstNmosvideoreceiver))
-#define GST_NMOSVIDEORECEIVER_CLASS(klass)                                                                             \
-    (G_TYPE_CHECK_CLASS_CAST((klass), GST_TYPE_NMOSVIDEORECEIVER, GstNmosvideoreceiverClass))
-#define GST_IS_NMOSVIDEORECEIVER(obj) (G_TYPE_CHECK_INSTANCE_TYPE((obj), GST_TYPE_NMOSVIDEORECEIVER))
-#define GST_IS_NMOSVIDEORECEIVER_CLASS(klass) (G_TYPE_CHECK_CLASS_TYPE((klass), GST_TYPE_NMOSVIDEORECEIVER))
+GST_DEBUG_CATEGORY_STATIC(gst_nmosaudioreceiver_debug_category);
+#define GST_CAT_DEFAULT gst_nmosaudioreceiver_debug_category
+#define GST_TYPE_NMOSAUDIORECEIVER (gst_nmosaudioreceiver_get_type())
+#define GST_NMOSAUDIORECEIVER(obj) (G_TYPE_CHECK_INSTANCE_CAST((obj), GST_TYPE_NMOSAUDIORECEIVER, GstNmosaudioreceiver))
+#define GST_NMOSAUDIORECEIVER_CLASS(klass)                                                                             \
+    (G_TYPE_CHECK_CLASS_CAST((klass), GST_TYPE_NMOSAUDIORECEIVER, GstNmosaudioreceiverClass))
+#define GST_IS_NMOSAUDIORECEIVER(obj) (G_TYPE_CHECK_INSTANCE_TYPE((obj), GST_TYPE_NMOSAUDIORECEIVER))
+#define GST_IS_NMOSAUDIORECEIVER_CLASS(klass) (G_TYPE_CHECK_CLASS_TYPE((klass), GST_TYPE_NMOSAUDIORECEIVER))
 
-typedef struct _GstNmosvideoreceiver
+typedef struct _GstNmosaudioreceiver
 {
     GstBin parent;
     GstPad* element_pad;
@@ -50,7 +52,8 @@ typedef struct _GstNmosvideoreceiver
     GstClock* clock;
     GstElementHandle<_GstElement> bin;
     GstElementHandle<_GstElement> udp_src;
-    GstElementHandle<_GstElement> rtp_video_depay;
+    GstElementHandle<_GstElement> rtp_audio_depay_16;
+    GstElementHandle<_GstElement> rtp_audio_depay_24;
     GstElementHandle<_GstElement> rtp_jitter_buffer;
     GstElementHandle<_GstElement> queue;
     ossrf::nmos_client_uptr client;
@@ -58,24 +61,21 @@ typedef struct _GstNmosvideoreceiver
     std::string sdp_string;
     sdp_settings_t sdp_settings;
     bool nmos_active;
-} GstNmosvideoreceiver;
+} GstNmosaudioreceiver;
 
-typedef struct _GstNmosvideoreceiverClass
+typedef struct _GstNmosaudioreceiverClass
 {
     GstBinClass parent_class;
-} GstNmosvideoreceiverClass;
+} GstNmosaudioreceiverClass;
 
-G_DEFINE_TYPE_WITH_CODE(GstNmosvideoreceiver, gst_nmosvideoreceiver, GST_TYPE_BIN,
-                        GST_DEBUG_CATEGORY_INIT(gst_nmosvideoreceiver_debug_category, "nmosvideoreceiver", 0,
+G_DEFINE_TYPE_WITH_CODE(GstNmosaudioreceiver, gst_nmosaudioreceiver, GST_TYPE_BIN,
+                        GST_DEBUG_CATEGORY_INIT(gst_nmosaudioreceiver_debug_category, "nmosaudioreceiver", 0,
                                                 "NMOS Receiver Plugin"))
 // Pad template
-static GstStaticPadTemplate src_template =
-    GST_STATIC_PAD_TEMPLATE("src", GST_PAD_SRC, GST_PAD_ALWAYS,
-                            GST_STATIC_CAPS("video/x-raw, "
-                                            "format=(string){ I420, NV12, RGB, YUV422P, UYVP }; "
-                                            "application/x-rtp, "
-                                            "media=(string)video, "
-                                            "clock-rate=(int)90000"));
+static GstStaticPadTemplate src_template = GST_STATIC_PAD_TEMPLATE("src", GST_PAD_SRC, GST_PAD_ALWAYS,
+                                                                   GST_STATIC_CAPS("audio/x-raw, "
+                                                                                   "format=(string){ S16BE, S24BE }; "
+                                                                                   "application/x-rtp"));
 
 enum class PropertyId : uint32_t
 {
@@ -91,10 +91,10 @@ enum class PropertyId : uint32_t
 };
 
 // Set properties so element variables can change depending on them
-static void gst_nmosvideoreceiver_set_property(GObject* object, guint property_id, const GValue* value,
+static void gst_nmosaudioreceiver_set_property(GObject* object, guint property_id, const GValue* value,
                                                GParamSpec* pspec)
 {
-    GstNmosvideoreceiver* self = GST_NMOSVIDEORECEIVER(object);
+    GstNmosaudioreceiver* self = GST_NMOSAUDIORECEIVER(object);
 
     switch(static_cast<PropertyId>(property_id))
     {
@@ -122,9 +122,9 @@ static void gst_nmosvideoreceiver_set_property(GObject* object, guint property_i
     }
 }
 
-static void gst_nmosvideoreceiver_get_property(GObject* object, guint property_id, GValue* value, GParamSpec* pspec)
+static void gst_nmosaudioreceiver_get_property(GObject* object, guint property_id, GValue* value, GParamSpec* pspec)
 {
-    GstNmosvideoreceiver* self = GST_NMOSVIDEORECEIVER(object);
+    GstNmosaudioreceiver* self = GST_NMOSAUDIORECEIVER(object);
 
     switch(static_cast<PropertyId>(property_id))
     {
@@ -150,63 +150,62 @@ static GstPadProbeReturn block_pad_probe_cb(GstPad* pad, GstPadProbeInfo* info, 
     return GST_PAD_PROBE_DROP;
 }
 
-void remove_old_bin(GstNmosvideoreceiver* self)
+void remove_old_bin(GstNmosaudioreceiver* self)
 {
-    if(self->element_pad != nullptr)
+    if(self->element_pad == nullptr)
     {
-        gulong block_id =
-            gst_pad_add_probe(self->element_pad, GST_PAD_PROBE_TYPE_BLOCK_DOWNSTREAM, block_pad_probe_cb, NULL, NULL);
+        return;
+    }
+    gulong block_id =
+        gst_pad_add_probe(self->element_pad, GST_PAD_PROBE_TYPE_BLOCK_DOWNSTREAM, block_pad_probe_cb, NULL, NULL);
 
-        gst_element_send_event(GST_ELEMENT(self), gst_event_new_flush_start());
-        gst_element_send_event(GST_ELEMENT(self), gst_event_new_flush_stop(FALSE));
+    gst_element_send_event(GST_ELEMENT(self), gst_event_new_flush_start());
+    gst_element_send_event(GST_ELEMENT(self), gst_event_new_flush_stop(FALSE));
 
-        self->clock = gst_element_get_clock(GST_ELEMENT(self));
+    self->clock = gst_element_get_clock(GST_ELEMENT(self));
 
-        gst_element_set_state(self->bin.get(), GST_STATE_NULL);
+    gst_element_set_state(self->bin.get(), GST_STATE_NULL);
 
-        gst_bin_remove(GST_BIN(self), self->udp_src.get());
-        gst_bin_remove(GST_BIN(self), self->rtp_jitter_buffer.get());
-        gst_bin_remove(GST_BIN(self), self->rtp_video_depay.get());
-        gst_bin_remove(GST_BIN(self), self->queue.get());
+    gst_bin_remove(GST_BIN(self), self->udp_src.get());
+    gst_bin_remove(GST_BIN(self), self->rtp_jitter_buffer.get());
+    gst_bin_remove(GST_BIN(self), self->rtp_audio_depay_16.get());
+    gst_bin_remove(GST_BIN(self), self->rtp_audio_depay_24.get());
+    gst_bin_remove(GST_BIN(self), self->queue.get());
 
-        gst_bin_remove(GST_BIN(self), self->bin.get());
+    gst_bin_remove(GST_BIN(self), self->bin.get());
 
-        self->udp_src.reset();
-        self->queue.reset();
-        self->rtp_jitter_buffer.reset();
-        self->rtp_video_depay.reset();
-        self->bin.reset();
+    self->udp_src.forget();
+    self->queue.forget();
+    self->rtp_jitter_buffer.forget();
+    self->rtp_audio_depay_16.forget();
+    self->rtp_audio_depay_24.forget();
+    self->bin.forget();
 
-        if(block_id != 0)
-        {
-            gst_pad_remove_probe(self->element_pad, block_id);
-        }
+    if(block_id != 0)
+    {
+        gst_pad_remove_probe(self->element_pad, block_id);
     }
 }
 
-void construct_pipeline(GstNmosvideoreceiver* self)
+void construct_pipeline(GstNmosaudioreceiver* self)
 {
     auto format = self->sdp_settings.format;
     try
     {
-        bisect::nmoscpp::video_sender_info_t video_info = std::get<bisect::nmoscpp::video_sender_info_t>(format);
-        auto& fr                                        = video_info.exact_framerate;
-        g_print("\n===== SDP Settings =====\n"
-                "Video (from variant):\n"
-                "  - height: %d\n"
-                "  - width:  %d\n"
-                "  - framerate: %ld/%ld\n"
-                "  - chroma_sub_sampling: %s\n"
-                "  - structure (interlace mode): %s\n"
-                "  - depth: %d\n"
-                "Primary leg:\n"
-                "  - destination_ip:   %s\n"
-                "  - destination_port: %d\n"
-                "\n",
-                video_info.height, video_info.width, fr.numerator(), fr.denominator(),
-                video_info.chroma_sub_sampling.c_str(), video_info.structure.name.c_str(), video_info.depth,
-                self->sdp_settings.primary.destination_ip.value().c_str(),
-                self->sdp_settings.primary.destination_port.value());
+        bisect::nmoscpp::audio_sender_info_t audio_info = std::get<bisect::nmoscpp::audio_sender_info_t>(format);
+        fmt::print("\n===== SDP Settings =====\n"
+                   "Audio (from variant):\n"
+                   "  - bits per sample: {}\n"
+                   "  - number of channels:  {}\n"
+                   "  - packet time: {}\n"
+                   "  - sampling rate: {}\n"
+                   "Primary leg:\n"
+                   "  - destination_ip:   {}\n"
+                   "  - destination_port: {}\n"
+                   "\n",
+                   audio_info.bits_per_sample, audio_info.number_of_channels, audio_info.packet_time,
+                   audio_info.sampling_rate, self->sdp_settings.primary.destination_ip.value().c_str(),
+                   self->sdp_settings.primary.destination_port.value());
 
         auto maybeBin = GstElementHandle<GstElement>::create_bin("dynamic-bin");
         if(std::holds_alternative<std::nullptr_t>(maybeBin))
@@ -215,74 +214,87 @@ void construct_pipeline(GstNmosvideoreceiver* self)
             return;
         }
 
-        GstElementHandle<GstElement> bin = std::move(std::get<GstElementHandle<GstElement>>(maybeBin));
+        self->bin = std::move(std::get<GstElementHandle<GstElement>>(maybeBin));
 
-        self->bin = std::move(bin);
-
-        auto maybeUdpsrc = GstElementHandle<GstElement>::create_element("udpsrc", nullptr);
-        auto maybeQueue  = GstElementHandle<GstElement>::create_element("queue", nullptr);
-        auto maybeJitter = GstElementHandle<GstElement>::create_element("rtpjitterbuffer", nullptr);
-        auto maybeDepay  = GstElementHandle<GstElement>::create_element("rtpvrawdepay", nullptr);
+        auto maybeUdpsrc  = GstElementHandle<GstElement>::create_element("udpsrc", nullptr);
+        auto maybeQueue   = GstElementHandle<GstElement>::create_element("queue", nullptr);
+        auto maybeJitter  = GstElementHandle<GstElement>::create_element("rtpjitterbuffer", nullptr);
+        auto maybeDepay16 = GstElementHandle<GstElement>::create_element("rtpL16depay", nullptr);
+        auto maybeDepay24 = GstElementHandle<GstElement>::create_element("rtpL24depay", nullptr);
 
         if(std::holds_alternative<std::nullptr_t>(maybeUdpsrc) || std::holds_alternative<std::nullptr_t>(maybeQueue) ||
-           std::holds_alternative<std::nullptr_t>(maybeJitter) || std::holds_alternative<std::nullptr_t>(maybeDepay))
+           std::holds_alternative<std::nullptr_t>(maybeJitter) ||
+           std::holds_alternative<std::nullptr_t>(maybeDepay16) || std::holds_alternative<std::nullptr_t>(maybeDepay24))
         {
             GST_ERROR_OBJECT(self, "Failed to create pipeline elements.");
             return;
         }
 
-        GstElementHandle<GstElement> udpsrc = std::move(std::get<GstElementHandle<GstElement>>(maybeUdpsrc));
-        GstElementHandle<GstElement> queue  = std::move(std::get<GstElementHandle<GstElement>>(maybeQueue));
-        GstElementHandle<GstElement> jitter = std::move(std::get<GstElementHandle<GstElement>>(maybeJitter));
-        GstElementHandle<GstElement> depay  = std::move(std::get<GstElementHandle<GstElement>>(maybeDepay));
+        self->udp_src            = std::move(std::get<GstElementHandle<GstElement>>(maybeUdpsrc));
+        self->queue              = std::move(std::get<GstElementHandle<GstElement>>(maybeQueue));
+        self->rtp_jitter_buffer  = std::move(std::get<GstElementHandle<GstElement>>(maybeJitter));
+        self->rtp_audio_depay_16 = std::move(std::get<GstElementHandle<GstElement>>(maybeDepay16));
+        self->rtp_audio_depay_24 = std::move(std::get<GstElementHandle<GstElement>>(maybeDepay24));
 
-        self->udp_src           = std::move(udpsrc);
-        self->queue             = std::move(queue);
-        self->rtp_jitter_buffer = std::move(jitter);
-        self->rtp_video_depay   = std::move(depay);
+        g_object_set(
+            G_OBJECT(self->udp_src.get()), "address", self->sdp_settings.primary.destination_ip.value().c_str(), "port",
+            self->sdp_settings.primary.destination_port.value(), "do-timestamp", true, "buffer-size", 212992, nullptr);
 
-        g_object_set(G_OBJECT(self->udp_src.get()), "address",
-                     self->sdp_settings.primary.destination_ip.value().c_str(), "port",
-                     self->sdp_settings.primary.destination_port.value(), "buffer-size", 67108864, nullptr);
-
-        // g_signal_connect(self->identity, "handoff", G_CALLBACK(identity_handoff_callback), NULL);
-
-        GstCaps* caps = gst_caps_new_simple(
-            "application/x-rtp", "media", G_TYPE_STRING, "video", "clock-rate", G_TYPE_INT, 90000, "encoding-name",
-            G_TYPE_STRING, "RAW", "payload", G_TYPE_INT, 96, "width", G_TYPE_STRING,
-            std::to_string(video_info.width).c_str(), "height", G_TYPE_STRING,
-            std::to_string(video_info.height).c_str(), "format", G_TYPE_STRING, "UYVP", "sampling", G_TYPE_STRING,
-            video_info.chroma_sub_sampling.c_str(), "depth", G_TYPE_STRING, std::to_string(video_info.depth).c_str(),
-            "framerate", GST_TYPE_FRACTION, fr.numerator(), fr.denominator(), nullptr);
+        GstCaps* caps = gst_caps_new_simple("application/x-rtp", "clock-rate", G_TYPE_INT, audio_info.sampling_rate,
+                                            "channels", G_TYPE_INT, audio_info.number_of_channels, nullptr);
 
         g_object_set(G_OBJECT(self->udp_src.get()), "caps", caps, nullptr);
 
-        g_object_set(G_OBJECT(self->rtp_jitter_buffer.get()), "do-retransmission", true, "do-lost", true, "mode", 2,
+        g_object_set(G_OBJECT(self->rtp_jitter_buffer.get()), "do-retransmission", true, "do-lost", true, "mode", 0,
                      "latency", 0, nullptr);
 
         g_object_set(G_OBJECT(self->queue.get()), "max-size-buffers", 3, nullptr);
 
         gst_bin_add_many(GST_BIN(self->bin.get()), self->udp_src.get(), self->rtp_jitter_buffer.get(),
-                         self->queue.get(), self->rtp_video_depay.get(), nullptr);
+                         self->queue.get(), self->rtp_audio_depay_16.get(), self->rtp_audio_depay_24.get(), nullptr);
 
         gst_element_sync_state_with_parent(self->udp_src.get());
         gst_element_sync_state_with_parent(self->rtp_jitter_buffer.get());
         gst_element_sync_state_with_parent(self->queue.get());
-        gst_element_sync_state_with_parent(self->rtp_video_depay.get());
+        gst_element_sync_state_with_parent(self->rtp_audio_depay_16.get());
+        gst_element_sync_state_with_parent(self->rtp_audio_depay_24.get());
 
-        if(gst_element_link_many(self->udp_src.get(), self->rtp_jitter_buffer.get(), self->queue.get(),
-                                 self->rtp_video_depay.get(), nullptr) == false)
+        GstPad* bin_src_pad = nullptr;
+
+        if(audio_info.bits_per_sample == 16)
         {
-            GST_ERROR_OBJECT(self, "Failed to link elements inside the bin.");
-            return;
+
+            if(gst_element_link_many(self->udp_src.get(), self->rtp_jitter_buffer.get(), self->queue.get(),
+                                     self->rtp_audio_depay_16.get(), nullptr) == false)
+            {
+                GST_ERROR_OBJECT(self, "Failed to link elements inside the bin.");
+                return;
+            }
+
+            // Create a ghost pad from the depay's src
+            bin_src_pad = gst_element_get_static_pad(self->rtp_audio_depay_16.get(), "src");
+            if(bin_src_pad == nullptr)
+            {
+                GST_ERROR_OBJECT(self, "Failed to get src pad from rtp_audio_depay.");
+                return;
+            }
         }
-
-        // Create a ghost pad from the depay's src
-        GstPad* bin_src_pad = gst_element_get_static_pad(self->rtp_video_depay.get(), "src");
-        if(bin_src_pad == nullptr)
+        else
         {
-            GST_ERROR_OBJECT(self, "Failed to get src pad from rtp_video_depay.");
-            return;
+            if(gst_element_link_many(self->udp_src.get(), self->rtp_jitter_buffer.get(), self->queue.get(),
+                                     self->rtp_audio_depay_24.get(), nullptr) == false)
+            {
+                GST_ERROR_OBJECT(self, "Failed to link elements inside the bin.");
+                return;
+            }
+
+            // Create a ghost pad from the depay's src
+            bin_src_pad = gst_element_get_static_pad(self->rtp_audio_depay_24.get(), "src");
+            if(bin_src_pad == nullptr)
+            {
+                GST_ERROR_OBJECT(self, "Failed to get src pad from rtp_audio_depay.");
+                return;
+            }
         }
 
         GstPad* bin_ghost_pad = gst_ghost_pad_new("src", bin_src_pad);
@@ -316,11 +328,11 @@ void construct_pipeline(GstNmosvideoreceiver* self)
     }
     catch(std::bad_variant_access const& ex)
     {
-        g_print("Invalid format sent for current receiver.\n");
+        fmt::print("Invalid format sent for current receiver.\n");
     }
 }
 
-void create_nmos(GstNmosvideoreceiver* self)
+void create_nmos(GstNmosaudioreceiver* self)
 {
     const auto node_config_json = create_node_config(self->config);
     if(node_config_json == nullptr)
@@ -333,12 +345,7 @@ void create_nmos(GstNmosvideoreceiver* self)
 
     receiver_config_json = create_receiver_config(self->config);
 
-    // g_print("Node Configuration: %s\n", node_config_json.dump().c_str());
-    // g_print("Device Configuration: %s\n", device_config_json.dump().c_str());
-    // g_print("Receiver Configuration: %s\n", receiver_config_json.dump().c_str());
-
     auto receiver_callback = [self](const std::optional<std::string>& sdp, bool master_enabled) {
-        // Debug print
         fmt::print("Receiver Activation Callback: SDP={}, Master Enabled={}\n", sdp.has_value() ? sdp.value() : "None",
                    master_enabled);
 
@@ -352,9 +359,7 @@ void create_nmos(GstNmosvideoreceiver* self)
                 self->sdp_string   = sdp.value();
                 remove_old_bin(self);
                 construct_pipeline(self);
-                auto time = gst_element_get_base_time(GST_ELEMENT(self));
                 gst_element_set_state(GST_ELEMENT(self), GST_STATE_PLAYING);
-                gst_pad_set_offset(self->bin_pad, (long)(gst_clock_get_time(self->clock) - time));
             }
         }
         else
@@ -365,23 +370,21 @@ void create_nmos(GstNmosvideoreceiver* self)
     };
 
     auto result = ossrf::nmos_client_t::create(self->config.node.id, node_config_json.dump());
-    if(result.has_value())
-    {
-        self->client = std::move(result.value());
-        self->client->add_device(device_config_json.dump());
-        self->client->add_receiver(self->config.device.id, receiver_config_json.dump(), receiver_callback);
-        GST_INFO_OBJECT(self, "NMOS client initialized successfully.");
-    }
-    else
+    if(!result.has_value())
     {
         GST_ERROR_OBJECT(self, "Failed to initialize NMOS client.");
+        return;
     }
+    self->client = std::move(result.value());
+    self->client->add_device(device_config_json.dump());
+    self->client->add_receiver(self->config.device.id, receiver_config_json.dump(), receiver_callback);
+    GST_INFO_OBJECT(self, "NMOS client initialized successfully.");
 }
 
-static GstStateChangeReturn gst_nmosvideoreceiver_change_state(GstElement* element, GstStateChange transition)
+static GstStateChangeReturn gst_nmosaudioreceiver_change_state(GstElement* element, GstStateChange transition)
 {
     GstStateChangeReturn ret;
-    GstNmosvideoreceiver* self = GST_NMOSVIDEORECEIVER(element);
+    GstNmosaudioreceiver* self = GST_NMOSAUDIORECEIVER(element);
 
     switch(transition)
     {
@@ -397,7 +400,7 @@ static GstStateChangeReturn gst_nmosvideoreceiver_change_state(GstElement* eleme
     default: break;
     }
 
-    ret = GST_ELEMENT_CLASS(gst_nmosvideoreceiver_parent_class)->change_state(element, transition);
+    ret = GST_ELEMENT_CLASS(gst_nmosaudioreceiver_parent_class)->change_state(element, transition);
 
     return ret;
 }
@@ -411,15 +414,15 @@ static void add_property(GObjectClass* object_class, guint id, const gchar* name
 }
 
 // Class initialization
-static void gst_nmosvideoreceiver_class_init(GstNmosvideoreceiverClass* klass)
+static void gst_nmosaudioreceiver_class_init(GstNmosaudioreceiverClass* klass)
 {
     GObjectClass* object_class     = G_OBJECT_CLASS(klass);
     GstElementClass* element_class = GST_ELEMENT_CLASS(klass);
 
     gst_element_class_add_pad_template(element_class, gst_static_pad_template_get(&src_template));
 
-    object_class->set_property = gst_nmosvideoreceiver_set_property;
-    object_class->get_property = gst_nmosvideoreceiver_get_property;
+    object_class->set_property = gst_nmosaudioreceiver_set_property;
+    object_class->get_property = gst_nmosaudioreceiver_get_property;
 
     add_property(object_class, 1, "node-id", "Node ID", "The ID of the node", "d49c85db-1c33-4f21-b160-58edd2af1810");
     add_property(
@@ -434,40 +437,40 @@ static void gst_nmosvideoreceiver_class_init(GstNmosvideoreceiverClass* klass)
     add_property(object_class, 6, "receiver-id", "Receiver ID", "UUID of the receiver",
                  "db9f46cf-2414-4e25-b6c6-2078159857f9");
     add_property(object_class, 7, "receiver-label", "Label of the receiver", "Label of the receiver",
-                 "BISECT OSSRF Video Receiver");
+                 "BISECT OSSRF Audio Receiver");
     add_property(object_class, 8, "receiver-description", "Description of the receiver", "Description of the receiver",
-                 "BISECT OSSRF Video Receiver");
+                 "BISECT OSSRF Audio Receiver");
     add_property(object_class, 9, "destination-address", "Destination Address", "Address of the destination",
                  "127.0.0.1");
 
-    gst_element_class_set_static_metadata(element_class, "NMOS Video Receiver", "Source/Network",
-                                          "Receives raw video from NMOS", "Luis Ferreira <luis.ferreira@bisect.pt>");
+    gst_element_class_set_static_metadata(element_class, "NMOS Audio Receiver", "Source/Network",
+                                          "Receives raw audio from NMOS", "Luis Ferreira <luis.ferreira@bisect.pt>");
 
-    element_class->change_state = gst_nmosvideoreceiver_change_state;
+    element_class->change_state = gst_nmosaudioreceiver_change_state;
 }
 
 // Object initialization
-static void gst_nmosvideoreceiver_init(GstNmosvideoreceiver* self)
+static void gst_nmosaudioreceiver_init(GstNmosaudioreceiver* self)
 {
     GstPadTemplate* src_tmpl = gst_static_pad_template_get(&src_template);
     GstPad* ghost_src        = gst_ghost_pad_new_no_target_from_template("src", src_tmpl);
     gst_object_unref(src_tmpl);
     gst_element_add_pad(GST_ELEMENT(self), ghost_src);
 
-    create_default_config_fields(&self->config);
+    create_default_config_fields_audio_receiver(&self->config);
 }
 
 static gboolean plugin_init(GstPlugin* plugin)
 {
-    return gst_element_register(plugin, "nmosvideoreceiver", GST_RANK_NONE, GST_TYPE_NMOSVIDEORECEIVER);
+    return gst_element_register(plugin, "nmosaudioreceiver", GST_RANK_NONE, GST_TYPE_NMOSAUDIORECEIVER);
 }
 
 #define VERSION "0.1"
-#define PACKAGE "gst-nmos-video-receiver-plugin"
+#define PACKAGE "gst-nmos-audio-receiver-plugin"
 #define PACKAGE_NAME "AMWA NMOS Sender and Receiver Framework Plugins"
 #define GST_PACKAGE_ORIGIN "https://www.amwa.tv/"
 #define GST_PACKAGE_LICENSE "Apache-2.0"
 
-GST_PLUGIN_DEFINE(GST_VERSION_MAJOR, GST_VERSION_MINOR, nmosvideoreceiver,
-                  "Plugin to receive video stream from NMOS application", plugin_init, VERSION, GST_PACKAGE_LICENSE,
+GST_PLUGIN_DEFINE(GST_VERSION_MAJOR, GST_VERSION_MINOR, nmosaudioreceiver,
+                  "Plugin to receive audio stream from NMOS application", plugin_init, VERSION, GST_PACKAGE_LICENSE,
                   PACKAGE_NAME, GST_PACKAGE_ORIGIN)
